app_web/pyproject.toml
[tool.poetry]
name = "app-web"
version = "1.75.0"
description = ""
authors = ["Alex Girlea"]

[tool.poetry.dependencies]
python = "~3.10"

Django = "~4.2"
mysql-connector-python = "~8.1.0"
gunicorn = "~21.2.0"
redis = "~5.0.0"
requests = "~2.31.0"
httpx = "~0.27.2"
rapidfuzz = "~3.10.1"

stripe = "~6.3.0"
celery = "~5.3.4"
flower = "~2.0.1"
cryptography = "~42.0.4"

# email / phone validation
email-validator = "~2.2.0"
boto3 = "~1.35.64"
phonenumbers = "~8.13.54"
twilio = "~9.4.4"

google-cloud-logging = "~3.6.0"
google-cloud-secret-manager = "~2.20.0"
google-cloud-recaptcha-enterprise = "~1.22.1"

# auth and user management
firebase-admin = "~6.2.0"
# oAuth will be kept at an exact version to avoid breaking changes in the DB
django-oauth-toolkit = "3.0.1"

# opentelemetry
opentelemetry-api = "~1.24.0"
opentelemetry-exporter-gcp-monitoring = "~1.6.0a0"
opentelemetry-exporter-gcp-trace = "~1.6.0"
opentelemetry-exporter-otlp-proto-common = "~1.24.0"
opentelemetry-exporter-otlp-proto-grpc = "~1.24.0"
opentelemetry-instrumentation = "~0.45b0"
opentelemetry-instrumentation-django = "~0.45b0"
opentelemetry-instrumentation-celery = "~0.45b0"
opentelemetry-instrumentation-wsgi = "~0.45b0"
opentelemetry-proto = "~1.24.0"
opentelemetry-resourcedetector-gcp = "~1.6.0a0"
opentelemetry-sdk = "~1.24.0"
opentelemetry-semantic-conventions = "~0.45b0"
opentelemetry-util-http = "~0.45b0"

[tool.poetry.group.dev.dependencies]
pytest = "~8.2.1"

[tool.poetry.group.container.dependencies]
mysqlclient = "~2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
